+++ fabric/src/main/java/net/mehvahdjukaar/polytone/fabric/PlatStuffImpl.java
@@ -191,6 +191,7 @@ public class PlatStuffImpl {
     }
 
     public static void sortTabs() {
+        CreativeModeTabs.validate();
     }
 
     public static CreativeModeTab registerTab(ResourceLocation id) {