+++ changelog.md
@@ -1,3 +1,1 @@
-- polytone can now convert optifine fog0, fog1 and fog2 and sky0, sky1 and sky2 to the Dimension Effect Modifiers colormaps
-- Creative tab icons can now accept a single item id as string
-- bumped major version number to 2.0 just so manual download people dont inevitably download this for the wrong version
+- Fixed custom creative tabs not sorting themselves properlty